:recycle: | Refactoring App.js to Routes
